+++ examples/chrome-extension/setup.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+curl -o vendor/honeybadger.min.js https://js.honeybadger.io/v3.2/honeybadger.min.js