+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased][latest]
 ### Added
 - Nodejs: Include source snippet in backtraces when available (#624)
+- `notifyAsync`: Async version of `notify` that returns a promise (#327)
 
 ### Changed
 - Call afterNotify handlers with error if notify preconditions fail (#654)