+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Call beforeNotify handlers even if preconditions fail (#654)
 - `Honeybadger.lambdaHandler`: return async or callback based handler based on input handler (#677)
 - Remove deprecated `disabled` config option (#671)
+- Apply `enableUncaught` setting to timers and event listeners (#690)
 
 ## [3.2.7] - 2021-11-01
 ### Fixed